backend/requirements.txt
# ─── Web Framework ──────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1       # ASGI server that runs FastAPI

# ─── Database ───────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.30     # Async ORM
asyncpg==0.29.0                 # Async PostgreSQL driver
alembic==1.13.1                 # Database migrations
pgvector==0.3.2                 # pgvector Python client (vector column type + query helpers)

# ─── Auth0 JWT Validation ────────────────────────────────────────
# python-jose: validates Auth0 JWTs (signature, expiry, audience, issuer)
# cryptography: needed by python-jose to verify RS256 (RSA) signatures
python-jose[cryptography]==3.3.0
httpx==0.27.0                   # Async HTTP client (fetch Auth0 JWKS public keys)

# ─── Kafka ───────────────────────────────────────────────────────
# confluent-kafka is the official, high-performance Kafka Python client
# maintained by Confluent (the company behind Kafka).
confluent-kafka==2.4.0

# ─── Pydantic / Settings ────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.3.1        # Reads config from .env file automatically

# ─── LLM Providers ──────────────────────────────────────────────
anthropic==0.28.0
openai==1.35.0

# ─── Email Alerts ───────────────────────────────────────────────
sendgrid==6.11.0

# ─── Utilities ──────────────────────────────────────────────────
python-dotenv==1.0.1
numpy==1.26.4                   # Needed for embedding operations with pgvector
